+++ R/add_ggsurvplot.R
@@ -3,6 +3,8 @@
 #'  object of class ggsurv, which is a list of ggplots.
 #'@param e1 an object of class ggsurv.
 #'@param e2 a plot component such as theme and labs.
+#'@note For ggplot2 v >= 3.5.2 compatibility, use \code{\%++\%} operator
+#'  instead of \code{+} when adding themes to ggsurv objects.
 #'@seealso \code{\link{theme_survminer}} and \code{\link{ggsurvplot}}
 #'@examples
 #'# Fit survival curves
@@ -18,7 +20,7 @@
 #'p
 #'
 #'# Customizing the plots
-#'p + theme_survminer(
+#'p %++% theme_survminer(
 #'      font.main = c(16, "bold", "darkblue"),
 #'      font.submain = c(15, "bold.italic", "purple"),
 #'      font.caption = c(14, "plain", "orange"),